+++ cmd/serverless/main.go
@@ -29,7 +29,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/serverless/proxy"
 	"github.com/DataDog/datadog-agent/pkg/serverless/random"
 	"github.com/DataDog/datadog-agent/pkg/serverless/registration"
-	"github.com/DataDog/datadog-agent/pkg/serverless/startchecker"
 	"github.com/DataDog/datadog-agent/pkg/serverless/trace"
 	"github.com/DataDog/datadog-agent/pkg/serverless/trace/inferredspan"
 	"github.com/DataDog/datadog-agent/pkg/trace/pb"
@@ -118,10 +117,8 @@ func runAgent(stopCh chan struct{}) (serverlessDaemon *daemon.Daemon, err error)
 
 	outputDatadogEnvVariablesForDebugging()
 
-	startChecker := startchecker.InitStartChecker()
-	startChecker.AddRule(&startchecker.ApiKeyEnvRule{})
-	if !startChecker.Check() {
-		log.Errorf("Can't start the Datadog extension: %s", startChecker.GetErrorMessage())
+	if !hasApiKey() {
+		log.Errorf("Can't start the Datadog extension as no API has been detected")
 		// we still need to register the extension but let's return after (no-op)
 		id, registrationError := registration.RegisterExtension(os.Getenv(runtimeAPIEnvVar), extensionRegistrationRoute, extensionRegistrationTimeout)
 		if registrationError != nil {
